+++ src/Sql/Sql/Database Backup/Model/AzureSqlDatabaseLongTermRetentionBackupModel.cs
@@ -78,6 +78,11 @@ namespace Microsoft.Azure.Commands.Sql.Backup.Model
         public string BackupStorageRedundancy { get; set; }
 
         /// <summary>
+        /// Gets or sets the setting whether the LTR backup is immutable
+        /// </summary>
+        public bool? IsBackupImmutable { get; set; }
+
+        /// <summary>
         /// Gets or sets the current backup storage access tier for the database
         /// </summary>
         public string BackupStorageAccessTier { get; set; }